requirements.txt
git+https://github.com/openai/whisper.git
torch
torchaudio
openai
python-dotenv
numpy
flask
flask-socketio
eventlet
